refactor(aih): consolidate list editor components into shared directory and make more generic
