+++ apps/ai-hero/src/app/_components/list-editor/dynamic-title.tsx
