+++ apps/ai-hero/src/app/admin/pages/_components/search-config.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { useSelection } from '@/app/(content)/lists/[slug]/edit/_components/selection-context'
+import { useSelection } from '@/app/_components/list-editor/selection-context'
 import { Configure } from 'react-instantsearch'
 
 export default function SearchConfig() {
